auth_service/requirements.txt
Django==4.1
djangorestframework==3.14
djangorestframework-simplejwt==5.2
psycopg2-binary==2.9.6
